Generalise for arbitrary comparators
While the Wu, et al scheme is described in terms of a greater-than comparison (to provide for ordering),
the scheme itself is general for any comparison function that returns a boolean result.

This change maintains backwards-compatibility with existing publicly-available code,
whilst providing the generalised interface necessary to support other comparison functions.