Implement constant-time ORE comparison
The subtle-ng crate does most of the heavy-lifting.

Note: we make use of the ConditionallySelectable trait, which does not
provide an implementation for usize, nor can we implement one here
because we own neither the trait definition, nor usize.

To work around it, I use 64 instead and convert back to usize when
required.